null check, count back to long, scratch comment

+++ src/main/java/dao/UserDao.java
@@ -62,14 +62,15 @@ public class UserDao {
                 .getSingleResult();
     }
 
+    // JPA count returns type Long
     public boolean userExists(String username) {
-        return (int) em.createQuery("SELECT COUNT(u.USER_ID) FROM User u WHERE u.username = :username")
+        return (Long) em.createQuery("SELECT COUNT(u.USER_ID) FROM User u WHERE u.username = :username")
                 .setParameter("username", username)
                 .getSingleResult() > 0;
     }
 
     public boolean emailExists(String email) {
-        return (int) em.createQuery("SELECT COUNT(u.USER_ID) FROM User u WHERE u.email = :email")
+        return (Long) em.createQuery("SELECT COUNT(u.USER_ID) FROM User u WHERE u.email = :email")
                 .setParameter("email", email)
                 .getSingleResult() > 0;
     }
